+++ spec/dummy/client/entries/app-kitten.js
@@ -7,6 +7,7 @@ import DocLinkBox from 'kitten/components/box/doc-link-box'
 import LinkBox from 'kitten/components/box/link-box'
 
 // Buttons
+import { Button } from 'kitten/components/buttons/button'
 import ButtonImageWithTextAndBadge from 'kitten/components/buttons/button-image-with-text-and-badge'
 import ButtonTooltipIcon from 'kitten/components/buttons/button-tooltip-icon'
 
@@ -33,6 +34,7 @@ import { Grid, GridCol } from 'kitten/components/grid/grid'
 
 // Icons
 import GrabberIcon from 'kitten/components/icons/grabber-icon'
+import { HeartIcon } from 'kitten/components/icons/heart-icon'
 
 // Links
 import ExternalRichLink from 'kitten/components/links/external-rich-link'
@@ -59,6 +61,7 @@ import TourStep from 'kitten/components/tours/tour-step'
 
 // Typography
 import { Paragraph } from 'kitten/components/typography/paragraph'
+import { Title } from 'kitten/components/typography/title'
 
 // Karl composed components
 
@@ -93,7 +96,9 @@ import { KarlParagraph,
          KarlParagraphQuaternary,
          KarlParagraphWithLink,
        } from 'kitten/karl/typography/paragraph'
-
+import { KarlButtonIconHeartRight,
+         KarlButtonIconHeartLeft
+       } from 'kitten/karl/buttons/button-heart-icon-words'
 
 ReactOnRails.register({
   // Box
@@ -101,6 +106,7 @@ ReactOnRails.register({
   LinkBox,
 
   // Buttons
+  Button,
   ButtonImageWithTextAndBadge,
   ButtonTooltipIcon,
 
@@ -124,7 +130,6 @@ ReactOnRails.register({
   LoanSimulator,
   ReactTooltip,
   Slider,
-
   Search,
   Select,
   SelectWithState,
@@ -133,6 +138,7 @@ ReactOnRails.register({
 
   // Icons
   GrabberIcon,
+  HeartIcon,
 
   // Links
   ExternalRichLink,
@@ -159,8 +165,11 @@ ReactOnRails.register({
 
   // Typography
   Paragraph,
+  Title,
 
   // Karl composed components
+  KarlButtonIconHeartRight,
+  KarlButtonIconHeartLeft,
   KarlImageExternalRichLink,
   KarlSvgExternalRichLink,
   KarlSimpleDropdown,